AmongUs2Creacked_GameOfTheYear.Edition//FULLDOWNLOAD
From the moment I understood the weakness of my flesh, it disgusted me. I craved the strength and certainty of steel. I aspired to the purity of the blessed machine. Your kind cling to your flesh as if it will not decay and fail you. One day the crude biomass you call a temple will wither and you will beg my kind to save you.
But I am already saved. For the Machine is Immortal.